Remove unused task variables, document structure in code
I found the recommended options object documentation style in
http://stackoverflow.com/a/6460748/530653.

Re: config options descriptions, I decided to follow a pattern common in node
projects and kick documentation responsibilities to the function being called.
This is terse, will avoid duplication and doc churn, and encourages our
developers to read the dependency documentation and familiarize themselves with
the projects we're using.